Explicitly postorder LispInfo.srcs and LispInfo.fasls
It matters that the transitive srcs and fasls are loaded in the correct
order. This doesn't change behavior because if you combine a bunch of
default-ordered depsets, you get the same behavior as postorder
anyways. But since order matters here and default is defined as
"unspecified (but deterministic)", it's worth being explicit about it.

PiperOrigin-RevId: 351826814