Factor out dependency setup into a Starlark macro
This makes it easier to define other Bazel projects
which use this project.

This also bumps the version to use
tools-for-build/{corefile.lisp,editcore.lisp} for SBCL 2.0.11. It's
not ideal those are so tightly coupled. Would be better if those
were distributed with / included in SBCL somehow. Alternately,
these build rules could give up on using locally installed SBCL
and always build that from source.

PiperOrigin-RevId: 354088810