Use progress_message placeholders
This reduces memory consumption, since the strings can be stored once instead of once per target. This changes the exact behavior since it uses File.path in the output instead of File.short_path and no longer puts the count of inputs in LispCombineFASLs (it's the same as the number of srcs for that target, so I'm not sure that detail in the progress message was particularly useful). The order of the depset for the inputs passed to LispCompile is specified explicitly because "%{input}" is expanded to the path of the first input in iteration order, and the default depset iteration order is "unspecified".

PiperOrigin-RevId: 390181774