Don't have a separate action type for block-compilation
This still sets bazel.main:*current-source-file* for compilation in the case where it's processing one file at a time. Would be better to use *compile-file-pathname*, but SBCL does not bind that when raising undefined-function errors from sb-c:compile-files. (Which means that error handling in the multiple-files block compilation case is still not working correctly.)

PiperOrigin-RevId: 415388166